Semillero relacion Con actividades

+++ src/main/java/com/API/datos/repository/ActividadesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActividadesRepository extends JpaRepository<Actividad, Long> {
+public interface ActividadesRepository extends JpaRepository<Actividad, Integer> {
 
 }
